requirements.txt
torch
torchvision
timm==0.4.12
Pillow
blobfile
mypy
numpy
pytest
requests
einops
tensorboardX
scipy
ftfy
opencv-python
sentencepiece
pyarrow
torchmetrics==0.7.3
transformers
pycocotools
pycocoevalcap
torchscale==0.2.0
